requirements.txt
ccxt>=4.3.0
pandas>=1.5.0,<2.3.0
ta>=0.11.0
numpy>=1.24.0,<2.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
colorlog>=6.8.0
rich>=13.0.0
